fix: resolve all clippy warnings for CI compliance
Applied comprehensive clippy fixes across the codebase:

Library fixes:
- Remove unused imports in session_manager and tls_server tests
- Fix CompressionLevel::to_flate2 to take self by value (Copy type)
- Use !is_empty() instead of len() > 0 in compression
- Add allow annotation for necessary pattern matching in message_queue
- Simplify hash_one usage in reconnect jitter calculation
- Fix doc list item indentation in string.rs
- Remove needless borrows in async_server tcp() calls
- Replace compile-time assert! with const assertion in udp
- Simplify image test data size calculation (remove * 1)

Examples fixes:
- Fix dynamic_receiver: use iterator for matrix rows, remove unused header
- Fix message_queue: prefix unused loop variable with _
- Fix logging: remove unused Level and fmt imports
- Fix sensor_logger: use iterators for array printing
- Fix image_streaming: replace max().min() with clamp()
- Fix video_streaming: remove unnecessary cast in capacity calculation
- Fix async_communication/performance_test: remove needless addr borrows
- Fix error_handling: remove unused SyncIgtlClient import
- Fix tls_communication: use eprintln!() instead of eprintln!("")

Tests fixes:
- Remove unused SyncIgtlClient import from client_helpers
- Use enumerate() iterator pattern in cpp_compat null-padding check

All changes maintain existing functionality while satisfying clippy's
strict warning requirements (-D warnings) for CI/CD pipeline.

+++ examples/video_streaming.rs
@@ -299,7 +299,7 @@ fn generate_h264_frame(frame_num: usize, width: u16, height: u16) -> Vec<u8> {
 ///
 /// Creates uncompressed RGB24 pixel data
 fn generate_raw_frame(frame_num: usize, width: u16, height: u16) -> Vec<u8> {
-    let mut data = Vec::with_capacity((width as usize * height as usize * 3) as usize);
+    let mut data = Vec::with_capacity(width as usize * height as usize * 3);
 
     for y in 0..height {
         for x in 0..width {
